Rename `intergrated_gui` to `gui`
related module imports are also updated
