+++ gui/component.py
