+++ gui/img/webcam.ico
