+++ gui/popup_shower.py
@@ -1,7 +1,7 @@
 from PyQt5.QtCore import QObject, pyqtSlot
 from PyQt5.QtWidgets import QApplication
 
-from intergrated_gui.popup_widget import TimeState, TimerWidget
+from gui.popup_widget import TimeState, TimerWidget
 
 
 class TimeShower(QObject):
